# Conflicts: (code format and same file update)
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/service/workrecord/erroralarm/recordcheck/ErAlWorkRecordCheckService.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.pubimp/src/main/java/nts/uk/ctx/at/record/pubimp/workrecord/erroralarm/recordcheck/ErAlWorkRecordCheckServicePubImpl.java
#	nts.uk/uk.at/at.ctx/schedule/nts.uk.ctx.at.schedule.pubimp/src/main/java/nts/uk/ctx/at/schedule/pubimp/schedule/basicschedule/ScBasicSchedulePubImpl.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/worktype/WorkTypeRepository.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/repository/worktype/JpaWorkTypeRepository.java
#	nts.uk/uk.com/wk.ctx/nts.uk.ctx.workflow.dom/src/main/java/nts/uk/ctx/workflow/dom/approverstatemanagement/ApprovalRootStateRepository.java
#	nts.uk/uk.com/wk.ctx/nts.uk.ctx.workflow.infra/src/main/java/nts/uk/ctx/workflow/infra/repository/approverstatemanagement/JpaApprovalRootStateRepository.java

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/service/detailscreen/before/BeforeProcessReleasingImpl.java
@@ -14,7 +14,7 @@ public class BeforeProcessReleasingImpl implements BeforeProcessReleasing {
 	private DetailBeforeUpdate detailBeforeProcessRegisterRepo;
 
 	@Override
-	public void detailScreenProcessBeforeReleasing(String companyId, String appId, Long version) {
+	public void detailScreenProcessBeforeReleasing(String companyId, String appId, int version) {
 		detailBeforeProcessRegisterRepo.exclusiveCheck(companyId, appId, version);		
 	}
 }
